Refactor Working Group page
