--- packages/ui/src/app/pages/WorkingGroups/Components.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const GroupContent = styled.div`
-  display: grid;
-  grid-row-gap: 24px;
-  width: 100%;
-`
