--- packages/ui/src/app/pages/WorkingGroups/WorkingGroup/index.ts
@@ -1,1 +1,0 @@
-export * from './WorkingGroup'
