--- packages/ui/src/app/pages/WorkingGroups/WorkingGroups/index.ts
@@ -1,1 +1,0 @@
-export * from './WorkingGroups'
